Introducing Robot Sensors (#52)
feat: Adds support for basic Contact and Distance sensors

This commit adds support for Contact (touch) and Distance sensors. These can be specified as part of the RobotSpec and assigned to appropriate channels. These readings can then be obtained via the RobotHandle using the `getDigitalInput()` and `getAnalogInput()` methods.

This is a squashed commit consisting of the following:
* Add specs and related interfaces to support sensors
* Add basic EventRegistry
* Basic Contact sensor implementation
* Adjust sensor transparency
* Contact sensor implementation
* Update code documentation
* Added Distance Sensor
* Fix linting

Resolves #46 
Resolves #21 
Resolves #47