fix: Fixes issue with robot not responding to motor commands
When a force is applied to the `SimRobotWheel` after some time of no
activity, the wheel body goes to sleep. the `applyForce()` method on it
has an optional 3rd parameter to wake the body up, which will then
actually apply the force.

Additionally, we also check and make sure that this is a sufficient
enough force to overcome static friction before moving.