SimRobot: Expose method to obtain the world rotation angle in radians
What prompted me to make this change is that we'd like to add a 'Gyroscope'
block to the Simulator.  Providing an interface on the SimRobot (and
RobotHandle) to get the robot's current world rotation angle  is the first
step towards this goal.

I choose to use the body's 'getAngle()' method according to the following
docs:
* https://box2d.org/documentation/md__d_1__git_hub_box2d_docs_dynamics.html#autotoc_md70
* https://box2d.org/documentation/classb2_body.html#ab5d135ef592b5f11f4e8d3ffbd831ac5