Add support for stepping outside rendering loop
In order to make the simulator deterministic we try to
link stepping through the code and running the world-sim.
The first step to do it is to make the world-sim be controlled
by the code execution.

This change add `step` function that will be called by the
code executor in specific places to advance the world.

This means the speed at which the code is executed will not
change how the simulation run.